Multiple Changes
New configuration parameters/flags for the config file:
- database path ("jdbc:h2:tcp://localhost/~/")
- database name (default: data type, e.g. "ssmi")
- generate images (default: false)

When you connect to an H2 database, now specify both the path and the
name.

Timestamp storage is re-enabled (I don't understand why that was
commented out)

Metadata storage is re-enabled

The "noCondensation" routine was reverted to the old version that
doesn't do any kind of condensation, just stores the data.

Various status messages and formatting changes